+++ playbook-website/app/helpers/application_helper.rb
@@ -133,55 +133,6 @@ module ApplicationHelper
     }
   end
 
-  def changelog_to_hash(changelog)
-    md_html = render_markdown(changelog)
-    document = Nokogiri::HTML5(md_html)
-    posts = []
-
-    # Loop Through Each Post
-    document.search("h1").each_with_index do |title, i|
-      index_element = i + 1
-      start_element = "//h1[#{index_element}]"
-      end_element = "//h1[#{index_element + 1}]"
-      set_a = "#{start_element}/following-sibling::*"
-      set_b = "#{end_element}/preceding-sibling::*"
-      my_content = document.xpath("#{set_a}[ count(.|#{set_b}) = count(#{set_b}) ]
-        | #{start_element} | #{end_element}")
-
-      # Extract attributes with defaults
-      image = my_content.search("img").map { |img| img["src"] }&.first
-      title_text = title.text
-
-      # Find the first non-empty paragraph <p> tag after the title or the image
-      description = my_content.xpath("#{start_element}/following-sibling::p[string-length(normalize-space(text())) > 0][1]")&.text || ""
-
-      date = my_content.search("h5").map(&:text)&.first
-      link = title.xpath("a[1]")[0]&.[]("href")&.gsub!("#", "")
-      content = my_content.search("p").to_s
-
-      # Create a hash with extracted attributes
-      post_hash = {
-        title: title_text,
-        description: description,
-        date: date,
-        image: image,
-        link: link,
-        content: content,
-      }
-
-      # Create the post object with OpenStruct
-      posts << OpenStruct.new(post_hash)
-    end
-
-    # Return Posts
-    posts
-  end
-
-  def changelog_to_object(changelog)
-    hash = changelog_to_hash(changelog)
-    hash.map { |post| OpenStruct.new(post) }
-  end
-
   def search_list
     all_kits = []
     formatted_kits = []
